test: drop redundant usize cast (clippy 1.96 unnecessary_cast)
diff_height is already usize; the cast tripped clippy::unnecessary_cast
on the CI toolchain (1.96), which is stricter than local 1.91.

+++ src/tui.rs
@@ -859,7 +859,7 @@ rename to src/relocated.rs
             24,
             &[key(KeyCode::Char('j')), key(KeyCode::Char('l'))],
         );
-        let half = a.diff_height as usize / 2;
+        let half = a.diff_height / 2;
         a.handle_key(ctrl('d'));
         assert_eq!(a.scroll, half.min(a.max_scroll()));
         a.handle_key(ctrl('u'));
